requirements.txt
streamlit==1.31.0
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
transformers==4.38.0
torch==2.2.0
sentence-transformers==2.5.0
spacy==3.7.4
ibm-watsonx-ai==0.2.6
ibm-watson==8.0.0
python-dotenv==1.0.1
pandas==2.2.0
nltk==3.8.1
regex==2023.12.25
langchain==0.1.10
langchain-ibm==0.1.0
huggingface-hub==0.21.0